refactor: MovePolicy 인터페이스 변경
확장 가능성을 고려하여, 인터페이스화 및 구현체로 분리
