+++ src/main/java/racingcar/policy/MovePolicy.java
@@ -1,0 +1,5 @@
+package racingcar.policy;
+
+public interface MovePolicy {
+    boolean canMove();
+}
