+++ src/main/java/racingcar/policy/RandomMovePolicy.java
@@ -1,8 +1,8 @@
-package racingcar.process;
+package racingcar.policy;
 
 import camp.nextstep.edu.missionutils.Randoms;
 
-public class MovePolicy {
+public class RandomMovePolicy implements MovePolicy {
     private final static Integer MIN_RANDOM_VALUE = 0;
     private final static Integer MAX_RANDOM_VALUE = 9;
     private final static Integer MOVE_THRESHOLD = 4;
