+++ src/test/java/racingcar/policy/RandomMovePolicyTest.java
@@ -1,4 +1,4 @@
-package racingcar.process;
+package racingcar.policy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -9,11 +9,12 @@ import java.util.List;
 import static camp.nextstep.edu.missionutils.test.Assertions.assertRandomNumberInRangeTest;
 import static org.junit.jupiter.api.Assertions.*;
 
-class MovePolicyTest {
+class RandomMovePolicyTest {
+    private static final MovePolicy policy = new RandomMovePolicy();
+
     @Test
     @DisplayName("랜덤 값에 따른 전진 테스트")
     void movePolicyTest() {
-        MovePolicy policy = new MovePolicy();
         List<Boolean> actual = new ArrayList<>();
         List<Boolean> expected = List.of(false, false, false, false, true, true, true, true, true, true);
